v0.16.0

+++ ffplayout-api/src/db/models.rs
@@ -4,7 +4,7 @@ use serde::{Deserialize, Serialize};
 pub struct User {
     #[sqlx(default)]
     #[serde(skip_deserializing)]
-    pub id: i64,
+    pub id: i32,
     #[sqlx(default)]
     pub mail: Option<String>,
     pub username: String,
@@ -16,10 +16,10 @@ pub struct User {
     pub salt: Option<String>,
     #[sqlx(default)]
     #[serde(skip_serializing)]
-    pub role_id: Option<i64>,
+    pub role_id: Option<i32>,
     #[sqlx(default)]
     #[serde(skip_serializing)]
-    pub channel_id: Option<i64>,
+    pub channel_id: Option<i32>,
     #[sqlx(default)]
     pub token: Option<String>,
 }
@@ -30,12 +30,12 @@ fn empty_string() -> String {
 
 #[derive(Debug, Deserialize, Serialize, Clone)]
 pub struct LoginUser {
-    pub id: i64,
+    pub id: i32,
     pub username: String,
 }
 
 impl LoginUser {
-    pub fn new(id: i64, username: String) -> Self {
+    pub fn new(id: i32, username: String) -> Self {
         Self { id, username }
     }
 }
@@ -43,8 +43,8 @@ impl LoginUser {
 pub struct TextPreset {
     #[sqlx(default)]
     #[serde(skip_deserializing)]
-    pub id: i64,
-    pub channel_id: i64,
+    pub id: i32,
+    pub channel_id: i32,
     pub name: String,
     pub text: String,
     pub x: String,
@@ -61,7 +61,7 @@ pub struct TextPreset {
 #[derive(Debug, Deserialize, Serialize, sqlx::FromRow)]
 pub struct Channel {
     #[serde(skip_deserializing)]
-    pub id: i64,
+    pub id: i32,
     pub name: String,
     pub preview_url: String,
     pub config_path: String,
